Initialise cursor in StaticMath blocks
Haydn and I went through the code to try to figure out why
the cursor is being rendered at all in StaticMath blocks.
It seems that the code replaces the focus blur event
handlers with just the initialisation code from the handler.

The problem is that they omitted the line which hides the
cursor in the initialisation, so it gets rendered anyway.

Here we just include the extra initialisation line which
resolves our issues.